feat: add form validation

+++ src/controller/form.php
@@ -4,22 +4,31 @@
         include_once './sendMessage.php';
 
         $data = [];
+        $msg = "Todos os campos devem ser preenchidos!";
 
         if(isset($_POST['title'])){
             $title = $_POST['title'];
             array_push($data, $title);
+        } else {
+            $blank = true;
         }
         if(isset($_POST['desc'])){
             $desc = $_POST['desc'];
             array_push($data, $desc);
+        }else {
+            $blank = true;
         }
         if(isset($_POST['price'])){
             $price = $_POST['price'];
             array_push($data, $price);
+        }else {
+            $blank = true;
         }
         if(isset($_POST['platform'])){
             $platform = $_POST['platform'];
             array_push($data, $platform);
+        }else {
+            $blank = true;
         }
         if(isset($_POST['played'])){
             $played = $_POST['played'];
@@ -29,6 +38,8 @@
                 array_push($data, false);
 
             }
+        }else {
+            $blank = true;
         }
         if(isset($_POST['releaseDate'])){
             $releaseDate = $_POST['releaseDate'];
@@ -36,47 +47,59 @@
             $dateSplit = explode("-", $releaseDate);
             $monthName = date("F", mktime(0, 0, 0, $dateSplit[1], 10));
             $date = $monthName . " "  . $dateSplit[2] . " " . $dateSplit[0];
+        }else {
+            $blank = true;
         }
 
 
         if(isset($_GET['id'])) {
-            $id = $_GET['id'];
-            $sql = "UPDATE game SET title = ?, description = ?, price = ?, platform = ?, played = ?, releaseDate = ? WHERE id = " . $id;
-            $conn->prepare($sql)->execute($data);
+            if ($blank == true) {
+                echo "a";
+                header("location: ../../public/pages/form.php?id=" . $id . "&msg=" . $msg);
+            } else {
+                $id = $_GET['id'];
+                $sql = "UPDATE game SET title = ?, description = ?, price = ?, platform = ?, played = ?, releaseDate = ? WHERE id = " . $id;
+                $conn->prepare($sql)->execute($data);
 
-            if(isset($_POST['category'])){
-                $conn->query("DELETE FROM gameCategory WHERE gameId = " . $id);
+                if(isset($_POST['category'])){
+                    $conn->query("DELETE FROM gameCategory WHERE gameId = " . $id);
 
-                foreach($_POST['category'] as $category ) {
-                    $sql = "INSERT INTO gameCategory (category, gameId) VALUES (?,?)";
-                    $conn->prepare($sql)->execute([$category, $id]);
+                    foreach($_POST['category'] as $category ) {
+                        $sql = "INSERT INTO gameCategory (category, gameId) VALUES (?,?)";
+                        $conn->prepare($sql)->execute([$category, $id]);
+                    }
                 }
+                header("location: ../../public/pages/list.php");
             }
-
         } else {
-            if(isset($_FILES['image'])){
-                $ext = strtolower(substr($_FILES['image']['name'],-4));
-                $imgName = date("Y.m.d-H.i.s") . $ext;
-                $dir = "../../public/uploads/";
-                move_uploaded_file($_FILES['image']['tmp_name'], $dir.$imgName);
-                array_push($data, $imgName);
-            }
-
-            $sql = "INSERT INTO game (title, description, price, platform, played, releaseDate, imageName) VALUES (?,?,?,?,?,?,?)";
-            $conn->prepare($sql)->execute($data);
-
-            $gameId = $conn->lastInsertId();
-
-            if(isset($_POST['category'])){
-                foreach($_POST['category'] as $category ) {
-                    $sql = "INSERT INTO gameCategory (category, gameId) VALUES (?,?)";
-                    $conn->prepare($sql)->execute([$category, $gameId]);
+            if ($blank == true) {
+                header("location: ../../public/pages/form.php?msg=" . $msg);
+                echo "a";
+            }else {
+                if(isset($_FILES['image'])){
+                    $ext = strtolower(substr($_FILES['image']['name'],-4));
+                    $imgName = date("Y.m.d-H.i.s") . $ext;
+                    $dir = "../../public/uploads/";
+                    move_uploaded_file($_FILES['image']['tmp_name'], $dir.$imgName);
+                    array_push($data, $imgName);
                 }
+
+                $sql = "INSERT INTO game (title, description, price, platform, played, releaseDate, imageName) VALUES (?,?,?,?,?,?,?)";
+                $conn->prepare($sql)->execute($data);
+
+                $gameId = $conn->lastInsertId();
+
+                if(isset($_POST['category'])){
+                    foreach($_POST['category'] as $category ) {
+                        $sql = "INSERT INTO gameCategory (category, gameId) VALUES (?,?)";
+                        $conn->prepare($sql)->execute([$category, $gameId]);
+                    }
+                }
+                header("location: ../../public/pages/list.php");
             }
         }
 
 
-        header("location: ../../public/pages/list.php");
     } catch(PDOException $e) {
         echo 'Ocorreu um erro durante a inserção: ' . $e->getMessage();
     }
